moved interfaces into the impl package
