+++ src/main/java/io/github/cottonmc/cotton_scripting/api/WorldStorage.java
@@ -1,13 +1,12 @@
 package io.github.cottonmc.cotton_scripting.api;
 
 import io.github.cottonmc.cotton_scripting.CottonScripting;
-import io.github.cottonmc.cotton_scripting.ExecutableScript;
+import io.github.cottonmc.cotton_scripting.impl.ExecutableScript;
 import io.github.cottonmc.cotton_scripting.impl.EntityWorldStorage;
 import io.github.cottonmc.cotton_scripting.impl.GlobalWorldStorage;
 import io.github.cottonmc.cotton_scripting.impl.ScriptTags;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.util.Identifier;
 
 import javax.annotation.Nullable;
