+++ src/main/java/io/github/cottonmc/cotton_scripting/event/ScriptedEventRunner.java
@@ -1,6 +1,6 @@
 package io.github.cottonmc.cotton_scripting.event;
 
-import io.github.cottonmc.cotton_scripting.ExecutableScript;
+import io.github.cottonmc.cotton_scripting.impl.ExecutableScript;
 import io.github.cottonmc.cotton_scripting.api.CottonScriptContext;
 import io.github.cottonmc.cotton_scripting.impl.ScriptTags;
 import io.github.cottonmc.functionapi.api.EventRunner;
