+++ src/main/java/io/github/cottonmc/cotton_scripting/impl/TagContainerAccessor.java
@@ -1,4 +1,4 @@
-package io.github.cottonmc.cotton_scripting;
+package io.github.cottonmc.cotton_scripting.impl;
 
 import net.minecraft.tag.Tag;
 import net.minecraft.util.Identifier;
